LPS-140172 Move Header to components folder and fix Deactivate behaviour with Key Details
