--- modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/containers/DXPActivationKeysTable/Header/index.js
@@ -1,190 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * The contents of this file are subject to the terms of the Liferay Enterprise
- * Subscription License ("License"). You may not use this file except in
- * compliance with the License. You can obtain a copy of the License by
- * contacting Liferay, Inc. See the License for the specific language governing
- * permissions and limitations under the License, including but not limited to
- * distribution rights of the Software.
- */
-
-import ClayAlert from '@clayui/alert';
-
-import classNames from 'classnames';
-import {useMemo, useState} from 'react';
-import {Button, ButtonDropDown} from '../../../../../common/components';
-import {ALERT_DOWNLOAD_TYPE} from '../../../utils/constants/alertDownloadType';
-import {AUTO_CLOSE_ALERT_TIME} from '../../../utils/constants/autoCloseAlertTime';
-import DeactivateButton from '../Deactivate';
-import {DOWNLOADABLE_LICENSE_KEYS} from '../utils/constants';
-import {ALERT_ACTIVATION_AGGREGATED_KEYS_DOWNLOAD_TEXT} from '../utils/constants/alertAggregateKeysDownloadText';
-import {getActivationKeyDownload} from '../utils/getActivationKeyDownload';
-import {getActivationKeysActionsItems} from '../utils/getActivationKeysActionsItems';
-import {getActivationKeysDownloadItems} from '../utils/getActivationKeysDownloadItems';
-
-const DXPActivationKeysTableHeader = ({
-	accountKey,
-	activationKeys,
-	licenseKeyDownloadURL,
-	selectedKeys,
-	sessionId,
-	setActivationKeys,
-}) => {
-	const [
-		activationKeysDownloadStatus,
-		setActivationKeysDownloadStatus,
-	] = useState('');
-
-	const selectedKeysIDs = selectedKeys
-		.map((selectedKey) => `licenseKeyIds=${selectedKey}`)
-		.join('&');
-
-	const isAbleToDownloadAggregateKeys = useMemo(() => {
-		const [
-			firstSelectedKey,
-			...restSelectedKeys
-		] = selectedKeys.map((key) =>
-			activationKeys.find((activationKey) => activationKey.id === key)
-		);
-
-		const keyCanBeDownloaded = restSelectedKeys.every(
-			(selectedKey) =>
-				DOWNLOADABLE_LICENSE_KEYS.above71DXPVersion(
-					firstSelectedKey,
-					selectedKey
-				) ||
-				DOWNLOADABLE_LICENSE_KEYS.below71DXPVersion(
-					firstSelectedKey,
-					selectedKey
-				)
-		);
-
-		return keyCanBeDownloaded;
-	}, [activationKeys, selectedKeys]);
-
-	const handleAlertStatus = (hasSuccessfullyDownloadedKeys) => {
-		setActivationKeysDownloadStatus(
-			hasSuccessfullyDownloadedKeys
-				? ALERT_DOWNLOAD_TYPE.success
-				: ALERT_DOWNLOAD_TYPE.danger
-		);
-	};
-
-	const activationKeysActionsItems = getActivationKeysActionsItems(
-		accountKey,
-		licenseKeyDownloadURL,
-		sessionId,
-		handleAlertStatus
-	);
-
-	const activationKeysDownloadItems = getActivationKeysDownloadItems(
-		isAbleToDownloadAggregateKeys,
-		selectedKeysIDs,
-		licenseKeyDownloadURL,
-		sessionId,
-		handleAlertStatus
-	);
-
-	const getCurrentButton = () => {
-		if (selectedKeys.length > 1) {
-			return (
-				<ButtonDropDown
-					items={activationKeysDownloadItems}
-					label="Download"
-					menuElementAttrs={{
-						className: 'p-0',
-					}}
-				/>
-			);
-		}
-
-		if (selectedKeys.length) {
-			return (
-				<Button
-					className="btn btn-primary"
-					onClick={async () =>
-						getActivationKeyDownload(
-							selectedKeys,
-							licenseKeyDownloadURL,
-							sessionId,
-							handleAlertStatus
-						)
-					}
-				>
-					Download
-				</Button>
-			);
-		}
-
-		return (
-			<ButtonDropDown
-				items={activationKeysActionsItems}
-				label="Actions"
-				menuElementAttrs={{
-					className: 'p-0',
-				}}
-			/>
-		);
-	};
-
-	return (
-		<div>
-			<div className="align-items-center bg-neutral-1 d-flex mb-2 p-3 rounded">
-				<div
-					className={classNames('align-items-center d-flex', {
-						'ml-auto': !selectedKeys.length,
-					})}
-				>
-					{!!selectedKeys.length && (
-						<>
-							<p className="font-weight-semi-bold m-0 ml-auto text-neutral-10">
-								{`${selectedKeys.length} Keys Selected`}
-							</p>
-
-							<DeactivateButton
-								selectedKeys={selectedKeys}
-								sessionId={sessionId}
-								setActivationKeys={setActivationKeys}
-							/>
-						</>
-					)}
-
-					{getCurrentButton()}
-				</div>
-			</div>
-
-			{activationKeysDownloadStatus && (
-				<ClayAlert.ToastContainer>
-					<ClayAlert
-						autoClose={
-							AUTO_CLOSE_ALERT_TIME[activationKeysDownloadStatus]
-						}
-						className="cp-activation-key-download-alert"
-						displayType={
-							ALERT_DOWNLOAD_TYPE[activationKeysDownloadStatus]
-						}
-						onClose={() => setActivationKeysDownloadStatus('')}
-					>
-						{
-							ALERT_ACTIVATION_AGGREGATED_KEYS_DOWNLOAD_TEXT[
-								activationKeysDownloadStatus
-							]
-						}
-					</ClayAlert>
-				</ClayAlert.ToastContainer>
-			)}
-
-			{!isAbleToDownloadAggregateKeys && (
-				<ClayAlert displayType="info">
-					To download an aggregate key, select keys with identical
-					<b>{' Type, Start Date, End Date, '}</b>
-					and
-					<b>{' Instance Size'} </b>
-				</ClayAlert>
-			)}
-		</div>
-	);
-};
-
-export default DXPActivationKeysTableHeader;
